adjust context structure
