+++ include/aquarius/router.hpp
@@ -6,9 +6,9 @@ namespace aquarius
 {
 	namespace detail
 	{
-		class context;
+		class basic_context;
 
-		using ctx_router = detail::single_router<std::shared_ptr<context>>;
+		using ctx_router = detail::single_router<std::shared_ptr<basic_context>>;
 
 		using msg_router = detail::single_router<std::shared_ptr<xmessage>>;
 	} // namespace detail
